pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vera-open"
version = "1.0.0"
description = "Morning OS for Notion users — AI-powered daily briefings via Telegram"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["notion", "telegram", "ai", "briefing", "productivity"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
]
dependencies = [
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "anthropic>=0.40.0",
]

[project.urls]
Homepage = "https://github.com/you/vera-open"
Issues = "https://github.com/you/vera-open/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
